IngestaOceanografica/requirements.txt
# Azure Functions core
azure-functions>=1.18.0

# Azure SDK
azure-eventhub>=5.11.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0

# HTTP requests
requests>=2.31.0

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0

# Utilidades
python-dateutil>=2.8.2
pytz>=2024.1
